requirements.txt
fastapi
httpx
python-jose[cryptography]
pytest
multidict
lru-dict